backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.35
aiosqlite==0.20.0
pydantic-settings==2.5.2
python-dotenv==1.0.1
google-generativeai==0.8.3
PyMuPDF==1.24.11
httpx==0.27.2
python-multipart==0.0.12
sse-starlette==2.1.3
faiss-cpu==1.8.0.post1
sentence-transformers==3.1.1
numpy==1.26.4
fpdf2==2.8.1
